+++ dev-projects/astro-content/src/consts.ts
@@ -1,0 +1,6 @@
+// Place any global data in this file.
+// You can import this data from anywhere in your site by using the `import` keyword.
+
+export const SITE_TITLE = 'Astro & Keystatic Blog';
+export const SITE_DESCRIPTION =
+  "A blog powered by Keystatic and Astro's content collections.";