+++ packages/keystatic/src/form/fields/text/index.tsx
@@ -43,6 +43,9 @@ export function validateText(
     if ((slugInfo.glob === '*' ? /[\\/]/ : /[\\]/).test(val)) {
       return `${fieldLabel} must not contain slashes`;
     }
+    if (/^\s|\s$/.test(val)) {
+      return `${fieldLabel} must not start or end with spaces`;
+    }
     if (slugInfo.slugs.has(val)) {
       return `${fieldLabel} must be unique`;
     }