+++ dev-projects/localization/CHANGELOG.md
@@ -1,5 +1,13 @@
 # localization
 
+## 0.1.87
+
+### Patch Changes
+
+- Updated dependencies [c43d7045]
+- Updated dependencies [21395048]
+  - @keystatic/core@0.0.116
+
 ## 0.1.86
 
 ### Patch Changes