+++ dev-projects/remix/CHANGELOG.md
@@ -1,5 +1,13 @@
 # @keystatic/remix-test-app
 
+## 0.0.79
+
+### Patch Changes
+
+- Updated dependencies [c43d7045]
+- Updated dependencies [21395048]
+  - @keystatic/core@0.0.116
+
 ## 0.0.78
 
 ### Patch Changes