+++ templates/astro/CHANGELOG.md
@@ -1,5 +1,16 @@
 # @keystatic/templates-astro
 
+## 0.0.8
+
+### Patch Changes
+
+- 2127a105: Add --host flag to astro dev command + version updates
+- Updated dependencies [5c86514f]
+- Updated dependencies [c43d7045]
+- Updated dependencies [21395048]
+  - @keystatic/astro@0.0.9
+  - @keystatic/core@0.0.116
+
 ## 0.0.7
 
 ### Patch Changes